# openapi.yaml
openapi: 3.1.0
info:
  title: authservicecentral HTTP API
  version: 1.0.0
  description: |
    The public HTTP contract for authservicecentral, a single-deployment
    authorization and OAuth 2.0 token-exchange service for human and workload
    identities.

    The service has three HTTP surfaces: OAuth discovery and token exchange,
    application authorization checks, and authenticated management operations.
    Management operations are grouped below `/v1/manage/` and require a
    platform JWT for the configured management audience. Each route requires
    the permission configured for its family and operation in the deployment
    YAML under `management.permissions`; when omitted, the conventional
    `management.<family>.<operation>` name is used. The examples below show the
    conventional names, not an unconditional hard-coded policy.
  contact:
    name: authservicecentral maintainers
  license:
    name: MIT
    identifier: MIT
servers:
  - url: "{scheme}://{host}:{port}"
    description: Deployment base URL
    variables:
      scheme:
        default: http
        enum: [http, https]
        description: Transport scheme used by the deployment.
      host:
        default: localhost
        description: Hostname where the service is reachable.
      port:
        default: "8080"
        description: Listening TCP port.
security: []
...
paths:
...
  /openapi.yaml:
    get:
      tags: [Documentation]
      operationId: getOpenAPISpecification
      summary: Return the embedded OpenAPI document
      description: Returns the canonical YAML document used by the interactive API reference.
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: OpenAPI 3.1 YAML document.
          content:
            application/yaml:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /.well-known/oauth-authorization-server:
    get:
      tags: [Discovery]
      operationId: getOAuthAuthorizationServerMetadata
      summary: Return OAuth authorization-server metadata
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: OAuth metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthMetadata"
        default:
          $ref: "#/components/responses/Error"
  /.well-known/jwks.json:
    get:
      tags: [Discovery]
      operationId: getJWKS
      summary: Return platform token verification keys
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: JSON Web Key Set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWKS"
        default:
          $ref: "#/components/responses/Error"
  /oauth2/token:
    post:
      tags: [OAuth]
      operationId: exchangeToken
      summary: Exchange a trusted JWT for an audience-scoped platform token
      description: |
        Implements OAuth 2.0 Token Exchange (RFC 8693). The subject token
        must be a JWT from a configured trusted token source. The requested
        audience must already exist. For delegated exchange, `actor_token` and
        `actor_token_type` must be supplied together and the audience's
        configured delegation policy must permit the request.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenExchangeRequest"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: Issued platform access token
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
            Pragma:
              schema: {type: string, const: no-cache}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenExchangeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/check:
    post:
      tags: [Authorization]
      operationId: checkPermissions
      summary: Batch-check resource permissions
      description: |
        The subject, audience, optional actor, delegation mode, and cached
        audience permissions are derived exclusively from the signed platform
        token. Caller-supplied identity fields are not accepted. A legitimate
        denial is returned as HTTP 200 with `allowed` set to false; malformed
        checks and checks for unknown or incompatible configured permissions
        are HTTP 400 errors.
      security:
        - platformBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckRequest"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: Ordered authorization decisions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Error"
  /v1/manage/audiences:
    get:
      tags: [Audiences]
      operationId: listAudiences
      summary: List audiences
      description: Requires the configured audiences.read management permission (by default `management.audiences.read`).
      security:
        - managementBearer: []
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: Audience list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudienceList"
        default:
          $ref: "#/components/responses/ManagementError"
    post:
      tags: [Audiences]
      operationId: putAudience
      summary: Create or update an audience
      description: Requires the configured audiences.write management permission (by default `management.audiences.write`). The audience ID is supplied in the body.
      security:
        - managementBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudienceInput"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "201":
          description: Stored audience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audience"
        default:
          $ref: "#/components/responses/ManagementError"
  /v1/manage/audiences/{id}:
    parameters:
      - $ref: "#/components/parameters/AudienceID"
    get:
      tags: [Audiences]
      operationId: getAudience
      summary: Get an audience
      description: Requires the configured audiences.read management permission (by default `management.audiences.read`).
      security:
        - managementBearer: []
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: Audience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audience"
        default:
          $ref: "#/components/responses/ManagementError"
    patch:
      tags: [Audiences]
      operationId: patchAudience
      summary: Partially update an audience
      description: Requires the configured audiences.write management permission (by default `management.audiences.write`).
      security:
        - managementBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudiencePatch"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: Updated audience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audience"
        default:
          $ref: "#/components/responses/ManagementError"
    delete:
      tags: [Audiences]
      operationId: deleteAudience
      summary: Delete an audience
      description: Requires the configured audiences.write management permission (by default `management.audiences.write`). Deletion is idempotent.
      security:
        - managementBearer: []
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "204":
          description: Audience deleted or already absent
        default:
          $ref: "#/components/responses/ManagementError"
  /v1/manage/resources:
    post:
      tags: [Resources]
      operationId: createResource
      summary: Create a resource
      description: Requires the configured resources.write management permission (by default `management.resources.write`).
      security:
        - managementBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Resource"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        default:
          $ref: "#/components/responses/ManagementError"
  /v1/manage/resources/{type}/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceType"
      - $ref: "#/components/parameters/ResourceID"
    get:
      tags: [Resources]
      operationId: getResource
      summary: Get a resource
      description: Requires the configured resources.read management permission (by default `management.resources.read`).
      security:
        - managementBearer: []
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: Resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        default:
          $ref: "#/components/responses/ManagementError"
    patch:
      tags: [Resources]
      operationId: patchResource
      summary: Replace resource metadata
      description: Requires the configured resources.write management permission (by default `management.resources.write`).
      security:
        - managementBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourcePatch"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        default:
          $ref: "#/components/responses/ManagementError"
    delete:
      tags: [Resources]
      operationId: deleteResource
      summary: Delete a resource and its authorization state
      description: Requires the configured resources.write management permission (by default `management.resources.write`). Deletion is idempotent.
      security:
        - managementBearer: []
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "204":
          description: Resource deleted or already absent
        default:
          $ref: "#/components/responses/ManagementError"
  /v1/manage/resources/{type}/{id}/relationships/{relation}:
    parameters:
      - $ref: "#/components/parameters/ResourceType"
      - $ref: "#/components/parameters/ResourceID"
      - $ref: "#/components/parameters/Relation"
    put:
      tags: [Resources]
      operationId: putResourceRelationship
      summary: Add or replace relationship targets
      description: Requires the configured resources.write management permission (by default `management.resources.write`). Cardinality is enforced by the configured schema.
      security:
        - managementBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipMutation"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "204":
          description: Relationship stored
        default:
          $ref: "#/components/responses/ManagementError"
    delete:
      tags: [Resources]
      operationId: deleteResourceRelationship
      summary: Delete relationship targets
      description: Requires the configured resources.write management permission (by default `management.resources.write`). With no body, all targets for the named relationship are removed.
      security:
        - managementBearer: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipMutation"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "204":
          description: Requested relationship targets removed
        default:
          $ref: "#/components/responses/ManagementError"
  /v1/manage/groups:
    post:
      tags: [Groups]
      operationId: createGroup
      summary: Create a group
      description: Requires the configured groups.write management permission (by default `management.groups.write`).
      security:
        - managementBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Group"
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        default:
          $ref: "#/components/responses/ManagementError"
...
  /v1/manage/grants/{id}:
    parameters:
      - $ref: "#/components/parameters/GrantID"
    delete:
      tags: [Grants]
      operationId: deleteGrant
      summary: Delete a grant
      description: Requires the configured grants.write management permission (by default `management.grants.write`). Deletion is idempotent.
      security:
        - managementBearer: []
      responses:
        "4XX":
          $ref: "#/components/responses/Error"
        "204":
          description: Grant deleted or already absent
        default:
          $ref: "#/components/responses/ManagementError"
components:
  securitySchemes:
    platformBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Platform JWT issued by this service for the requested audience. Required by `/v1/check`.
    managementBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Platform JWT whose `aud` is the configured management audience and whose `permissions` claim includes the operation-specific permission configured under `management.permissions` (or its conventional default).
...
  responses:
...
    ServiceUnavailable:
      description: A required dependency is unavailable or the service is not ready
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...